Refactor domain modules and test coverage
