+++ src/note_todo_sync/mod.rs
@@ -1,0 +1,13 @@
+pub mod checklist;
+pub mod metadata;
+pub mod sync;
+
+pub use checklist::{
+    checkbox_sync_enabled, parse_checklist_items, render_checklist_line, set_checkbox_sync_enabled,
+    upsert_mapping_token, ChecklistItem,
+};
+pub use metadata::{normalize_text, parse_metadata, PRIORITY_MAX, PRIORITY_MIN};
+pub use sync::{
+    sync_note_todos, PreviewAction, RevisionState, SyncConfig, SyncConflict, SyncMode, SyncResult,
+    TodoItem,
+};
