+++ src/settings/mod.rs
@@ -1,0 +1,7 @@
+pub mod defaults;
+pub mod io;
+pub mod model;
+
+pub use defaults::*;
+pub use io::{set_settings_path, settings_path};
+pub use model::*;
